+++ src/main/java/com/app/folioman/mfschemes/MFSchemeDTO.java
@@ -1,4 +1,4 @@
-package com.app.folioman.shared;
+package com.app.folioman.mfschemes;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
