+++ src/main/java/com/app/folioman/mfschemes/mapper/MfSchemeDtoToEntityMapper.java
@@ -1,7 +1,7 @@
 package com.app.folioman.mfschemes.mapper;
 
+import com.app.folioman.mfschemes.MFSchemeDTO;
 import com.app.folioman.mfschemes.entities.MfFundScheme;
-import com.app.folioman.shared.MFSchemeDTO;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
