+++ src/main/java/com/app/folioman/mfschemes/mapper/MfSchemeDtoToEntityMapperHelper.java
@@ -1,5 +1,6 @@
 package com.app.folioman.mfschemes.mapper;
 
+import com.app.folioman.mfschemes.MFSchemeDTO;
 import com.app.folioman.mfschemes.entities.MFSchemeNav;
 import com.app.folioman.mfschemes.entities.MFSchemeType;
 import com.app.folioman.mfschemes.entities.MfAmc;
@@ -7,7 +8,6 @@ import com.app.folioman.mfschemes.entities.MfFundScheme;
 import com.app.folioman.mfschemes.service.MFSchemeTypeService;
 import com.app.folioman.mfschemes.service.MfAmcService;
 import com.app.folioman.mfschemes.util.SchemeConstants;
-import com.app.folioman.shared.MFSchemeDTO;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Arrays;
