+++ src/main/java/com/app/folioman/mfschemes/service/CachedNavService.java
@@ -1,7 +1,7 @@
 package com.app.folioman.mfschemes.service;
 
+import com.app.folioman.mfschemes.MFSchemeDTO;
 import com.app.folioman.mfschemes.NavNotFoundException;
-import com.app.folioman.shared.MFSchemeDTO;
 import java.time.LocalDate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
