+++ src/main/java/com/app/folioman/mfschemes/service/MFSchemeNavService.java
@@ -1,11 +1,11 @@
 package com.app.folioman.mfschemes.service;
 
+import com.app.folioman.mfschemes.MFNavService;
+import com.app.folioman.mfschemes.MFSchemeDTO;
 import com.app.folioman.mfschemes.NavNotFoundException;
 import com.app.folioman.mfschemes.repository.MFSchemeNavRepository;
 import com.app.folioman.mfschemes.util.SchemeConstants;
 import com.app.folioman.shared.LocalDateUtility;
-import com.app.folioman.shared.MFNavService;
-import com.app.folioman.shared.MFSchemeDTO;
 import java.time.LocalDate;
 import java.util.List;
 import org.slf4j.Logger;
