+++ src/main/java/com/app/folioman/mfschemes/service/MfHistoricalNavService.java
@@ -1,11 +1,11 @@
 package com.app.folioman.mfschemes.service;
 
+import com.app.folioman.mfschemes.MFSchemeDTO;
 import com.app.folioman.mfschemes.NavNotFoundException;
 import com.app.folioman.mfschemes.entities.MfFundScheme;
 import com.app.folioman.mfschemes.mapper.MfSchemeDtoToEntityMapper;
 import com.app.folioman.mfschemes.util.SchemeConstants;
 import com.app.folioman.shared.CommonConstants;
-import com.app.folioman.shared.MFSchemeDTO;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.Reader;
