+++ src/main/java/com/app/folioman/portfolio/service/UserSchemeDetailsService.java
@@ -1,9 +1,9 @@
 package com.app.folioman.portfolio.service;
 
+import com.app.folioman.mfschemes.MFNavService;
 import com.app.folioman.portfolio.entities.UserSchemeDetails;
 import com.app.folioman.portfolio.repository.UserSchemeDetailsRepository;
 import com.app.folioman.shared.FundDetailProjection;
-import com.app.folioman.shared.MFNavService;
 import com.app.folioman.shared.MFSchemeProjection;
 import com.app.folioman.shared.MfSchemeService;
 import com.app.folioman.shared.UserSchemeDetailService;
